Move RowBody component to the assets atoms folder
