+++ src/components/molecules/Spinner.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Loader from 'react-loader-spinner';
 import "react-loader-spinner/dist/loader/css/react-spinner-loader.css";
-import { RowBody } from '../atoms/RowBody';
+import { RowBody } from '../../assets/styles/atoms/RowBody';
 
 export default () => (
   <RowBody spacing='center'>
